write_event: use POST api/event background feature to support sqlite
With the parameter background, the event is written after the main tx has completed. This works for the specific case to write an event, but e.g. doesn't catch error during save etc. To fix that, we can add a feature to re-use an open tx inside the plugin, but that is for another day.

see #64339